+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 - Fixed handling of post releases.
 - Fixed python restricted dependencies not being checked agaisnt virtualenv version.
+- Fixed python/platform constraint not being picked up for subdependencies.
 
 
 ## [0.4.2] - 2018-03-10